Add documentation and major feature updates
Added multiple documentation files detailing architecture, corrections, features, and improvements. Introduced new services (AudioEffectsProcessor, ChordSuggestionEngine) and views (AudioEffectsSheet, ChordDiagramView, RecordingDetailView). Updated models and views to support audio effects, chord suggestions, improved UI/UX, and fixed switch exhaustiveness errors. Enhanced project and recording management, swipe actions, and audio processing features.

+++ Suonote/Models/ChordEvent.swift
@@ -5,8 +5,8 @@ import SwiftData
 final class ChordEvent {
     var id: UUID
     var barIndex: Int
-    var beatOffset: Int
-    var duration: Int
+    var beatOffset: Double  // Changed to Double to support fractional beats
+    var duration: Double    // Duration in beats (0.5 = half beat, 1 = full beat, 2 = two beats)
     var root: String
     var quality: ChordQuality
     var extensions: [String]
@@ -17,8 +17,8 @@ final class ChordEvent {
     
     init(
         barIndex: Int,
-        beatOffset: Int,
-        duration: Int = 1,
+        beatOffset: Double,
+        duration: Double = 1.0,
         root: String,
         quality: ChordQuality = .major,
         extensions: [String] = [],
